ds-scripts/requirements.txt
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anthropic==0.28.0
anyio==4.3.0
asgiref==3.8.1
assemblyai==0.26.0
astrapy==1.2.1
async-timeout==4.0.3
asyncer==0.0.5
attrs==23.2.0
Authlib==1.3.0
backoff==2.2.1
bce-python-sdk==0.9.9
bcrypt==4.0.1
beautifulsoup4==4.12.3
bidict==0.23.1
boto3==1.34.103
botocore==1.34.103
bson==0.5.10
build==1.2.1
cachetools==5.3.3
cassandra-driver==3.29.1
cassio==0.1.7
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
clevercsv==0.8.2
click==8.1.7
cohere==5.5.7
colorclass==2.2.2
coloredlogs==15.0.1
colorlog==6.8.2
compressed_rtf==1.0.6
cryptography==42.0.7
dataclasses-json==0.6.6
datasets==2.14.7
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.0.0
docstring-parser==0.15
dspy-ai==2.4.9
duckdb
easygui==0.98.3
ebcdic==1.1.1
ecdsa==0.19.0
elastic-transport==8.13.0
elasticsearch==8.13.1
email_validator==2.1.1
emoji==2.12.1
exceptiongroup==1.2.1
extract-msg==0.47.0
faiss-cpu==1.8.0
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.4
filelock==3.14.0
filetype==1.2.0
FLAML==2.1.2
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2023.10.0
future==1.0.0
geomet==0.2.1.post1
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.51.0
google-cloud-bigquery==3.22.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.2
google-resumable-media==2.7.0
google_search_results==2.4.2
googleapis-common-protos==1.63.0
gotrue==2.4.2
greenlet==3.0.3
groq==0.5.0
grpc-google-iam-v1==0.13.0
grpcio==1.64.0rc1
grpcio-health-checking==1.63.0rc1
grpcio-status==1.63.0rc1
grpcio-tools==1.63.0rc1
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
ijson==3.2.3
importlib-metadata==7.0.0
importlib_resources==6.4.0
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.3.2
jq==1.7.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.2.3
langchain-anthropic==0.1.15
langchain-astradb==0.3.3
langchain-chroma==0.1.1
langchain-cohere==0.1.7
langchain-community==0.2.4
langchain-core==0.2.5
langchain-experimental==0.0.60
langchain-google-genai==1.0.6
langchain-google-vertexai==1.0.5
langchain-groq==0.1.5
langchain-mistralai==0.1.8
langchain-openai==0.1.8
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.1
langchainhub==0.1.15
langdetect==1.0.9
langflow==1.0.0a53
langflow-base==0.0.64
langfuse==2.36.1
langsmith==0.1.77
lark==1.1.8
litellm==1.40.9
llama-index==0.10.36
llama-index-agent-openai==0.2.4
llama-index-cli==0.1.12
llama-index-core==0.10.36
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.18
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.22
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.2
llamaindex-py-client==0.1.19
loguru==0.7.2
lxml==5.2.2
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
metal_sdk==2.5.1
metaphor-python==0.1.23
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msoffcrypto-tool==5.4.0
multidict==6.0.5
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
oletools==0.60.1
onnxruntime==1.17.3
openai==1.28.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
optuna==3.6.1
ordered-set==4.1.0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parameterized==0.9.0
passlib==1.7.4
pcodedmp==1.2.6
pgvector==0.2.5
pillow==10.3.0
pinecone-client==3.2.2
platformdirs==4.2.1
portalocker==2.8.2
postgrest==0.16.4
posthog==3.5.0
prompt-toolkit==3.0.43
proto-plus==1.24.0.dev1
protobuf==4.25.3
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg2-binary==2.9.9
pulsar-client==3.5.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen==0.2.27
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
pymongo==4.7.2
pyparsing==2.4.7
pypdf==4.2.0
pyperclip==1.8.2
PyPika==0.48.9
pyproject_hooks==1.1.0
pysrt==1.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-iso639==2024.4.27
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.7
python-pptx==0.6.23
python-socketio==5.11.2
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.9.1
qianfan==0.3.5
rapidfuzz==3.9.3
realtime==1.0.4
red-black-tree-mod==1.20
regex==2024.5.10
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
RTFDE==0.1.1
s3transfer==0.10.1
shapely==2.0.4
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlmodel==0.0.18
starlette==0.37.2
storage3==0.7.4
StrEnum==0.4.15
striprtf==0.0.26
structlog==24.1.0
supabase==2.4.5
supafunc==0.4.5
sympy==1.12.1rc1
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tqdm==4.66.4
typer==0.12.3
types-cachetools==5.3.0.7
types-protobuf==4.25.0.20240417
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
uncurl==0.0.11
unstructured==0.14.5
unstructured-client==0.23.3
upstash-vector==0.4.0
uritemplate==4.1.1
urllib3==2.2.1
uuid6==2024.1.12
uvicorn==0.30.1
uvloop==0.19.0
validators==0.28.1
watchfiles==0.21.0
wcwidth==0.2.13
weaviate-client==4.6.0
websocket-client==1.8.0
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
youtube-transcript-api==0.6.2
zep-python==2.0.0rc6
zipp==3.18.1
